Combining DataTypeInt and DataTypeInt64
Make the Name fields the same so that they are both handled by IntField.

+++ pkg/generator/metadata/types.go
@@ -16,7 +16,7 @@ var (
 	DataTypeFloat32     = DataType{Name: "Decimal", Literal: "float32", NullableLiteral: "null.Float"}
 	DataTypeFloat64     = DataType{Name: "Decimal", Literal: "float64", NullableLiteral: "null.Float"}
 	DataTypeInt         = DataType{Name: "Int", Literal: "int", NullableLiteral: "null.Int"}
-	DataTypeInt64       = DataType{Name: "Int64", Literal: "int64", NullableLiteral: "null.Int"}
+	DataTypeInt64       = DataType{Name: "Int", Literal: "int64", NullableLiteral: "null.Int"}
 	DataTypeBigInt      = DataType{Name: "BigInt", Literal: "big.Int", NullableLiteral: "big.Int"}
 	DataTypeBigFloat    = DataType{Name: "BigFloat", Literal: "big.Float", NullableLiteral: "nullable.BigFloat"}
 	DataTypeJSONB       = DataType{Name: "Jsonb", Literal: "[]byte", NullableLiteral: "nullable.Jsonb"}
